maree/pyproject.toml
[project]
name = "maree"
version = "0.1.0"
description = "Pluggable document ingestion pipeline for RAG"
requires-python = ">=3.12"
dependencies = [
    "click>=8.0",
    "pydantic-settings>=2.0",
    "httpx>=0.27",
    "asyncpg>=0.30",
    "pgvector>=0.3",
    "pypdf>=5.0",
]

[project.optional-dependencies]
vertex = ["PyJWT[crypto]>=2.9.0"]
docling = ["docling>=2.0"]
dev = ["pytest>=8.0", "pytest-asyncio>=0.24"]

[project.scripts]
maree = "maree.cli:main"

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"
